+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/next/objects/asset/delegates/AssetModernDelegate.kt
@@ -22,6 +22,7 @@ import expo.modules.medialibrary.next.extensions.resolver.queryGetCreationTime
 import expo.modules.medialibrary.next.extensions.resolver.updateRelativePath
 import expo.modules.medialibrary.next.objects.wrappers.RelativePath
 import expo.modules.medialibrary.next.objects.asset.Asset
+import expo.modules.medialibrary.next.objects.wrappers.MediaType
 import expo.modules.medialibrary.next.objects.wrappers.MimeType
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -61,7 +62,7 @@ class AssetModernDelegate(override val contentUri: Uri, context: Context) : Asse
     val height = contentResolver.queryAssetHeight(contentUri)
       ?: throw AssetPropertyNotFoundException("Height")
     // If height is not saved to the database
-    if (getMediaType().contains("image") && height <= 0) {
+    if (getMediaType() == MediaType.IMAGE && height <= 0) {
       return downloadBitmapAndGet { it.outHeight }
     }
     return height
@@ -70,7 +71,7 @@ class AssetModernDelegate(override val contentUri: Uri, context: Context) : Asse
   override suspend fun getWidth(): Int {
     val width = contentResolver.queryAssetWidth(contentUri)
       ?: throw AssetPropertyNotFoundException("Width")
-    if (getMediaType().contains("image") && width <= 0) {
+    if (getMediaType() == MediaType.IMAGE && width <= 0) {
       return downloadBitmapAndGet { it.outWidth }
     }
     return width
@@ -83,9 +84,8 @@ class AssetModernDelegate(override val contentUri: Uri, context: Context) : Asse
     return extract(options)
   }
 
-  override suspend fun getMediaType(): String =
-    contentResolver.getType(contentUri)
-      ?: throw AssetPropertyNotFoundException("MediaType")
+  override suspend fun getMediaType(): MediaType =
+    MediaType.fromContentUri(contentUri)
 
   override suspend fun getModificationTime(): Long? =
     contentResolver.queryAssetModificationTime(contentUri).takeIf { it != 0L }