+++ packages/expo-media-library/src/next/types/MediaType.ts
@@ -1,0 +1,6 @@
+export enum MediaType {
+  UNKNOWN = 'unknown',
+  IMAGE = 'image',
+  AUDIO = 'audio',
+  VIDEO = 'video',
+}